+++ demo.py
@@ -41,5 +41,6 @@ mdl = model_fn()
 mdl.summary()
 # which metrics to track across the runs and display
 METRIC_TO_REPORT = [('loss', 'min'), ('val_loss', 'min'), ('acc', 'max'), ('val_acc', 'max')]
-hypermodel = RandomSearch(model_fn, epoch_budget=90, max_epochs=10, dry_run=DRY_RUN, project="kerastuner-demo", architecture="MLP", username=username, metrics=METRIC_TO_REPORT)
+hypermodel = RandomSearch(model_fn, epoch_budget=90, max_epochs=10, dry_run=DRY_RUN, project="kerastuner-demo", architecture="MLP", metrics=METRIC_TO_REPORT)
+hypermodel.backend(username=username)
 hypermodel.search(x_train, y_train, validation_split=0.01)